+++ fe/fe-core/src/main/java/org/apache/doris/nereids/trees/expressions/ExpressionEvaluator.java
@@ -21,11 +21,16 @@ import org.apache.doris.catalog.Env;
 import org.apache.doris.common.AnalysisException;
 import org.apache.doris.nereids.trees.expressions.functions.ExecutableFunctions;
 import org.apache.doris.nereids.trees.expressions.functions.agg.AggregateFunction;
+import org.apache.doris.nereids.trees.expressions.functions.executable.DateTimeArithmetic;
+import org.apache.doris.nereids.trees.expressions.functions.executable.NumericArithmetic;
+import org.apache.doris.nereids.trees.expressions.functions.executable.TimeRoundSeries;
+import org.apache.doris.nereids.trees.expressions.literal.DateLiteral;
 import org.apache.doris.nereids.trees.expressions.literal.Literal;
 import org.apache.doris.nereids.trees.expressions.literal.NullLiteral;
 import org.apache.doris.nereids.types.DataType;
 import org.apache.doris.nereids.types.DecimalV3Type;
 
+import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMultimap;
 
 import java.lang.reflect.InvocationTargetException;
@@ -123,15 +128,23 @@ public enum ExpressionEvaluator {
         }
         ImmutableMultimap.Builder<String, FunctionInvoker> mapBuilder =
                 new ImmutableMultimap.Builder<String, FunctionInvoker>();
-        Class clazz = ExecutableFunctions.class;
-        for (Method method : clazz.getDeclaredMethods()) {
-            ExecFunctionList annotationList = method.getAnnotation(ExecFunctionList.class);
-            if (annotationList != null) {
-                for (ExecFunction f : annotationList.value()) {
-                    registerFEFunction(mapBuilder, method, f);
+        List<Class> classes = ImmutableList.of(
+                ExecutableFunctions.class,
+                DateLiteral.class,
+                DateTimeArithmetic.class,
+                NumericArithmetic.class,
+                TimeRoundSeries.class
+        );
+        for (Class cls : classes) {
+            for (Method method : cls.getDeclaredMethods()) {
+                ExecFunctionList annotationList = method.getAnnotation(ExecFunctionList.class);
+                if (annotationList != null) {
+                    for (ExecFunction f : annotationList.value()) {
+                        registerFEFunction(mapBuilder, method, f);
+                    }
                 }
+                registerFEFunction(mapBuilder, method, method.getAnnotation(ExecFunction.class));
             }
-            registerFEFunction(mapBuilder, method, method.getAnnotation(ExecFunction.class));
         }
         this.functions = mapBuilder.build();
     }
