+++ fe/fe-core/src/main/java/org/apache/doris/nereids/trees/expressions/functions/ExecutableFunctions.java
@@ -30,17 +30,11 @@ import org.apache.doris.nereids.trees.expressions.literal.DoubleLiteral;
 import org.apache.doris.nereids.trees.expressions.literal.FloatLiteral;
 import org.apache.doris.nereids.trees.expressions.literal.IntegerLiteral;
 import org.apache.doris.nereids.trees.expressions.literal.LargeIntLiteral;
-import org.apache.doris.nereids.trees.expressions.literal.Literal;
-import org.apache.doris.nereids.trees.expressions.literal.NullLiteral;
 import org.apache.doris.nereids.trees.expressions.literal.SmallIntLiteral;
 import org.apache.doris.nereids.trees.expressions.literal.TinyIntLiteral;
 import org.apache.doris.nereids.trees.expressions.literal.VarcharLiteral;
-import org.apache.doris.nereids.types.DecimalV3Type;
 import org.apache.doris.nereids.util.DateUtils;
 
-import java.math.BigDecimal;
-import java.math.BigInteger;
-import java.time.Duration;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Calendar;
@@ -91,1930 +85,8 @@ public class ExecutableFunctions {
         SECOND
     }
 
-    /**
-     * Executable arithmetic functions add
-     */
-    @ExecFunction(name = "add", argTypes = {"TINYINT", "TINYINT"}, returnType = "SMALLINT")
-    public static SmallIntLiteral addTinyIntTinyInt(TinyIntLiteral first, TinyIntLiteral second) {
-        short result = (short) Math.addExact(first.getValue(), second.getValue());
-        return new SmallIntLiteral(result);
-    }
 
-    @ExecFunction(name = "add", argTypes = {"TINYINT", "SMALLINT"}, returnType = "INT")
-    public static IntegerLiteral addTinyIntSmallInt(TinyIntLiteral first, SmallIntLiteral second) {
-        int result = Math.addExact(first.getValue(), second.getValue());
-        return new IntegerLiteral(result);
-    }
 
-    @ExecFunction(name = "add", argTypes = {"TINYINT", "INT"}, returnType = "BIGINT")
-    public static BigIntLiteral addTinyIntInt(TinyIntLiteral first, IntegerLiteral second) {
-        long result = Math.addExact((long) first.getValue(), (long) second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"TINYINT", "BIGINT"}, returnType = "BIGINT")
-    public static BigIntLiteral addTinyIntBigInt(TinyIntLiteral first, BigIntLiteral second) {
-        long result = Math.addExact((long) first.getValue(), (long) second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"TINYINT", "LARGEINT"}, returnType = "LARGEINT")
-    public static LargeIntLiteral addTinyIntLargeInt(TinyIntLiteral first, LargeIntLiteral second) {
-        BigInteger result = second.getValue().add(new BigInteger(first.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"SMALLINT", "TINYINT"}, returnType = "INT")
-    public static IntegerLiteral addSmallIntTinyInt(SmallIntLiteral first, TinyIntLiteral second) {
-        int result = Math.addExact(first.getValue(), second.getValue());
-        return new IntegerLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"SMALLINT", "SMALLINT"}, returnType = "INT")
-    public static IntegerLiteral addSmallIntSmallInt(SmallIntLiteral first, SmallIntLiteral second) {
-        int result = Math.addExact(first.getValue(), second.getValue());
-        return new IntegerLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"SMALLINT", "INT"}, returnType = "BIGINT")
-    public static BigIntLiteral addSmallIntInt(SmallIntLiteral first, IntegerLiteral second) {
-        long result = Math.addExact((long) first.getValue(), (long) second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"SMALLINT", "BIGINT"}, returnType = "BIGINT")
-    public static BigIntLiteral addSmallIntBigInt(SmallIntLiteral first, BigIntLiteral second) {
-        long result = Math.addExact(first.getValue(), second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"SMALLINT", "LARGEINT"}, returnType = "LARGEINT")
-    public static LargeIntLiteral addSmallIntLargeInt(SmallIntLiteral first, LargeIntLiteral second) {
-        BigInteger result = second.getValue().add(new BigInteger(first.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"INT", "TINYINT"}, returnType = "BIGINT")
-    public static BigIntLiteral addIntTinyInt(IntegerLiteral first, TinyIntLiteral second) {
-        long result = Math.addExact((long) first.getValue(), (long) second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"INT", "SMALLINT"}, returnType = "BIGINT")
-    public static BigIntLiteral addIntSmallInt(IntegerLiteral first, SmallIntLiteral second) {
-        long result = Math.addExact((long) first.getValue(), (long) second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"INT", "INT"}, returnType = "BIGINT")
-    public static BigIntLiteral addIntInt(IntegerLiteral first, IntegerLiteral second) {
-        long result = Math.addExact((long) first.getValue(), (long) second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"INT", "BIGINT"}, returnType = "BIGINT")
-    public static BigIntLiteral addIntBigInt(IntegerLiteral first, BigIntLiteral second) {
-        long result = Math.addExact(first.getValue(), second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"INT", "LARGEINT"}, returnType = "LARGEINT")
-    public static LargeIntLiteral addIntLargeInt(IntegerLiteral first, LargeIntLiteral second) {
-        BigInteger result = second.getValue().add(new BigInteger(first.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"BIGINT", "TINYINT"}, returnType = "BIGINT")
-    public static BigIntLiteral addBigIntTinyInt(BigIntLiteral first, TinyIntLiteral second) {
-        long result = Math.addExact(first.getValue(), second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"BIGINT", "SMALLINT"}, returnType = "BIGINT")
-    public static BigIntLiteral addBigIntSmallInt(BigIntLiteral first, SmallIntLiteral second) {
-        long result = Math.addExact(first.getValue(), second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"BIGINT", "INT"}, returnType = "BIGINT")
-    public static BigIntLiteral addBigIntInt(BigIntLiteral first, IntegerLiteral second) {
-        long result = Math.addExact(first.getValue(), second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"BIGINT", "BIGINT"}, returnType = "BIGINT")
-    public static BigIntLiteral addBigIntBigInt(BigIntLiteral first, BigIntLiteral second) {
-        long result = Math.addExact(first.getValue(), second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"BIGINT", "LARGEINT"}, returnType = "LARGEINT")
-    public static LargeIntLiteral addBigIntLargeInt(BigIntLiteral first, LargeIntLiteral second) {
-        BigInteger result = second.getValue().add(new BigInteger(first.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"LARGEINT", "TINYINT"}, returnType = "BIGINT")
-    public static LargeIntLiteral addLargeIntTinyInt(LargeIntLiteral first, TinyIntLiteral second) {
-        BigInteger result = first.getValue().add(new BigInteger(second.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"LARGEINT", "SMALLINT"}, returnType = "BIGINT")
-    public static LargeIntLiteral addLargeIntSmallInt(LargeIntLiteral first, SmallIntLiteral second) {
-        BigInteger result = first.getValue().add(new BigInteger(second.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"LARGEINT", "INT"}, returnType = "BIGINT")
-    public static LargeIntLiteral addLargeIntInt(LargeIntLiteral first, IntegerLiteral second) {
-        BigInteger result = first.getValue().add(new BigInteger(second.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"LARGEINT", "BIGINT"}, returnType = "BIGINT")
-    public static LargeIntLiteral addLargeIntBigInt(LargeIntLiteral first, BigIntLiteral second) {
-        BigInteger result = first.getValue().add(new BigInteger(second.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"LARGEINT", "LARGEINT"}, returnType = "LARGEINT")
-    public static LargeIntLiteral addLargeIntLargeInt(LargeIntLiteral first, LargeIntLiteral second) {
-        BigInteger result = first.getValue().add(new BigInteger(second.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"DOUBLE", "DOUBLE"}, returnType = "DOUBLE")
-    public static DoubleLiteral addDoubleDouble(DoubleLiteral first, DoubleLiteral second) {
-        double result = first.getValue() + second.getValue();
-        return new DoubleLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"DECIMAL", "DECIMAL"}, returnType = "DECIMAL")
-    public static DecimalLiteral addDecimalDecimal(DecimalLiteral first, DecimalLiteral second) {
-        BigDecimal result = first.getValue().add(second.getValue());
-        return new DecimalLiteral(result);
-    }
-
-    @ExecFunction(name = "add", argTypes = {"DECIMALV3", "DECIMALV3"}, returnType = "DECIMALV3")
-    public static DecimalV3Literal addDecimalV3DecimalV3(DecimalV3Literal first, DecimalV3Literal second) {
-        BigDecimal result = first.getValue().add(second.getValue());
-        return new DecimalV3Literal((DecimalV3Type) first.getDataType(), result);
-    }
-
-    /**
-     * Executable arithmetic functions subtract
-     */
-    @ExecFunction(name = "subtract", argTypes = {"TINYINT", "TINYINT"}, returnType = "SMALLINT")
-    public static SmallIntLiteral subtractTinyIntTinyInt(TinyIntLiteral first, TinyIntLiteral second) {
-        short result = (short) Math.subtractExact(first.getValue(), second.getValue());
-        return new SmallIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"TINYINT", "SMALLINT"}, returnType = "INT")
-    public static IntegerLiteral subtractTinyIntSmallInt(TinyIntLiteral first, SmallIntLiteral second) {
-        int result = Math.subtractExact(first.getValue(), second.getValue());
-        return new IntegerLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"TINYINT", "INT"}, returnType = "BIGINT")
-    public static BigIntLiteral subtractTinyIntInt(TinyIntLiteral first, IntegerLiteral second) {
-        long result = Math.subtractExact((long) first.getValue(), (long) second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"TINYINT", "BIGINT"}, returnType = "BIGINT")
-    public static BigIntLiteral subtractTinyIntBigInt(TinyIntLiteral first, BigIntLiteral second) {
-        long result = Math.subtractExact((long) first.getValue(), (long) second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"TINYINT", "LARGEINT"}, returnType = "LARGEINT")
-    public static LargeIntLiteral subtractTinyIntLargeInt(TinyIntLiteral first, LargeIntLiteral second) {
-        BigInteger result = second.getValue().subtract(new BigInteger(first.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"SMALLINT", "TINYINT"}, returnType = "INT")
-    public static IntegerLiteral subtractSmallIntTinyInt(SmallIntLiteral first, TinyIntLiteral second) {
-        int result = Math.subtractExact(first.getValue(), second.getValue());
-        return new IntegerLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"SMALLINT", "SMALLINT"}, returnType = "INT")
-    public static IntegerLiteral subtractSmallIntSmallInt(SmallIntLiteral first, SmallIntLiteral second) {
-        int result = Math.subtractExact(first.getValue(), second.getValue());
-        return new IntegerLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"SMALLINT", "INT"}, returnType = "BIGINT")
-    public static BigIntLiteral subtractSmallIntInt(SmallIntLiteral first, IntegerLiteral second) {
-        long result = Math.subtractExact((long) first.getValue(), (long) second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"SMALLINT", "BIGINT"}, returnType = "BIGINT")
-    public static BigIntLiteral subtractSmallIntBigInt(SmallIntLiteral first, BigIntLiteral second) {
-        long result = Math.subtractExact(first.getValue(), second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"SMALLINT", "LARGEINT"}, returnType = "LARGEINT")
-    public static LargeIntLiteral subtractSmallIntLargeInt(SmallIntLiteral first, LargeIntLiteral second) {
-        BigInteger result = second.getValue().subtract(new BigInteger(first.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"INT", "TINYINT"}, returnType = "BIGINT")
-    public static BigIntLiteral subtractIntTinyInt(IntegerLiteral first, TinyIntLiteral second) {
-        long result = Math.subtractExact((long) first.getValue(), (long) second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"INT", "SMALLINT"}, returnType = "BIGINT")
-    public static BigIntLiteral subtractIntSmallInt(IntegerLiteral first, SmallIntLiteral second) {
-        long result = Math.subtractExact((long) first.getValue(), (long) second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"INT", "INT"}, returnType = "BIGINT")
-    public static BigIntLiteral subtractIntInt(IntegerLiteral first, IntegerLiteral second) {
-        long result = Math.subtractExact((long) first.getValue(), (long) second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"INT", "BIGINT"}, returnType = "BIGINT")
-    public static BigIntLiteral subtractIntBigInt(IntegerLiteral first, BigIntLiteral second) {
-        long result = Math.subtractExact(first.getValue(), second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"INT", "LARGEINT"}, returnType = "LARGEINT")
-    public static LargeIntLiteral subtractIntLargeInt(IntegerLiteral first, LargeIntLiteral second) {
-        BigInteger result = second.getValue().subtract(new BigInteger(first.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"BIGINT", "TINYINT"}, returnType = "BIGINT")
-    public static BigIntLiteral subtractBigIntTinyInt(BigIntLiteral first, TinyIntLiteral second) {
-        long result = Math.subtractExact(first.getValue(), second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"BIGINT", "SMALLINT"}, returnType = "BIGINT")
-    public static BigIntLiteral subtractBigIntSmallInt(BigIntLiteral first, SmallIntLiteral second) {
-        long result = Math.subtractExact(first.getValue(), second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"BIGINT", "INT"}, returnType = "BIGINT")
-    public static BigIntLiteral subtractBigIntInt(BigIntLiteral first, IntegerLiteral second) {
-        long result = Math.subtractExact(first.getValue(), second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"BIGINT", "BIGINT"}, returnType = "BIGINT")
-    public static BigIntLiteral subtractBigIntBigInt(BigIntLiteral first, BigIntLiteral second) {
-        long result = Math.subtractExact(first.getValue(), second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"BIGINT", "LARGEINT"}, returnType = "LARGEINT")
-    public static LargeIntLiteral subtractBigIntLargeInt(BigIntLiteral first, LargeIntLiteral second) {
-        BigInteger result = second.getValue().subtract(new BigInteger(first.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"LARGEINT", "TINYINT"}, returnType = "BIGINT")
-    public static LargeIntLiteral subtractLargeIntTinyInt(LargeIntLiteral first, TinyIntLiteral second) {
-        BigInteger result = first.getValue().subtract(new BigInteger(second.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"LARGEINT", "SMALLINT"}, returnType = "BIGINT")
-    public static LargeIntLiteral subtractLargeIntSmallInt(LargeIntLiteral first, SmallIntLiteral second) {
-        BigInteger result = first.getValue().subtract(new BigInteger(second.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"LARGEINT", "INT"}, returnType = "BIGINT")
-    public static LargeIntLiteral subtractLargeIntInt(LargeIntLiteral first, IntegerLiteral second) {
-        BigInteger result = first.getValue().subtract(new BigInteger(second.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"LARGEINT", "BIGINT"}, returnType = "BIGINT")
-    public static LargeIntLiteral subtractLargeIntBigInt(LargeIntLiteral first, BigIntLiteral second) {
-        BigInteger result = first.getValue().subtract(new BigInteger(second.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"LARGEINT", "LARGEINT"}, returnType = "LARGEINT")
-    public static LargeIntLiteral subtractLargeIntLargeInt(LargeIntLiteral first, LargeIntLiteral second) {
-        BigInteger result = first.getValue().subtract(new BigInteger(second.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"DOUBLE", "DOUBLE"}, returnType = "DOUBLE")
-    public static DoubleLiteral subtractDoubleDouble(DoubleLiteral first, DoubleLiteral second) {
-        double result = first.getValue() - second.getValue();
-        return new DoubleLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"DECIMAL", "DECIMAL"}, returnType = "DECIMAL")
-    public static DecimalLiteral subtractDecimalDecimal(DecimalLiteral first, DecimalLiteral second) {
-        BigDecimal result = first.getValue().subtract(second.getValue());
-        return new DecimalLiteral(result);
-    }
-
-    @ExecFunction(name = "subtract", argTypes = {"DECIMALV3", "DECIMALV3"}, returnType = "DECIMALV3")
-    public static DecimalV3Literal subtractDecimalV3DecimalV3(DecimalV3Literal first, DecimalV3Literal second) {
-        BigDecimal result = first.getValue().subtract(second.getValue());
-        return new DecimalV3Literal((DecimalV3Type) first.getDataType(), result);
-    }
-
-    /**
-     * Executable arithmetic functions multiply
-     */
-    @ExecFunction(name = "multiply", argTypes = {"TINYINT", "TINYINT"}, returnType = "SMALLINT")
-    public static SmallIntLiteral multiplyTinyIntTinyInt(TinyIntLiteral first, TinyIntLiteral second) {
-        short result = (short) Math.multiplyExact(first.getValue(), second.getValue());
-        return new SmallIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"TINYINT", "SMALLINT"}, returnType = "INT")
-    public static IntegerLiteral multiplyTinyIntSmallInt(TinyIntLiteral first, SmallIntLiteral second) {
-        int result = Math.multiplyExact(first.getValue(), second.getValue());
-        return new IntegerLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"TINYINT", "INT"}, returnType = "BIGINT")
-    public static BigIntLiteral multiplyTinyIntInt(TinyIntLiteral first, IntegerLiteral second) {
-        long result = Math.multiplyExact((long) first.getValue(), (long) second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"TINYINT", "BIGINT"}, returnType = "BIGINT")
-    public static BigIntLiteral multiplyTinyIntBigInt(TinyIntLiteral first, BigIntLiteral second) {
-        long result = Math.multiplyExact((long) first.getValue(), (long) second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"TINYINT", "LARGEINT"}, returnType = "LARGEINT")
-    public static LargeIntLiteral multiplyTinyIntLargeInt(TinyIntLiteral first, LargeIntLiteral second) {
-        BigInteger result = second.getValue().multiply(new BigInteger(first.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"SMALLINT", "TINYINT"}, returnType = "INT")
-    public static IntegerLiteral multiplySmallIntTinyInt(SmallIntLiteral first, TinyIntLiteral second) {
-        int result = Math.multiplyExact(first.getValue(), second.getValue());
-        return new IntegerLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"SMALLINT", "SMALLINT"}, returnType = "INT")
-    public static IntegerLiteral multiplySmallIntSmallInt(SmallIntLiteral first, SmallIntLiteral second) {
-        int result = Math.multiplyExact(first.getValue(), second.getValue());
-        return new IntegerLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"SMALLINT", "INT"}, returnType = "BIGINT")
-    public static BigIntLiteral multiplySmallIntInt(SmallIntLiteral first, IntegerLiteral second) {
-        long result = Math.multiplyExact((long) first.getValue(), (long) second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"SMALLINT", "BIGINT"}, returnType = "BIGINT")
-    public static BigIntLiteral multiplySmallIntBigInt(SmallIntLiteral first, BigIntLiteral second) {
-        long result = Math.multiplyExact(first.getValue(), second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"SMALLINT", "LARGEINT"}, returnType = "LARGEINT")
-    public static LargeIntLiteral multiplySmallIntLargeInt(SmallIntLiteral first, LargeIntLiteral second) {
-        BigInteger result = second.getValue().multiply(new BigInteger(first.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"INT", "TINYINT"}, returnType = "BIGINT")
-    public static BigIntLiteral multiplyIntTinyInt(IntegerLiteral first, TinyIntLiteral second) {
-        long result = Math.multiplyExact((long) first.getValue(), (long) second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"INT", "SMALLINT"}, returnType = "BIGINT")
-    public static BigIntLiteral multiplyIntSmallInt(IntegerLiteral first, SmallIntLiteral second) {
-        long result = Math.multiplyExact((long) first.getValue(), (long) second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"INT", "INT"}, returnType = "BIGINT")
-    public static BigIntLiteral multiplyIntInt(IntegerLiteral first, IntegerLiteral second) {
-        long result = Math.multiplyExact((long) first.getValue(), (long) second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"INT", "BIGINT"}, returnType = "BIGINT")
-    public static BigIntLiteral multiplyIntBigInt(IntegerLiteral first, BigIntLiteral second) {
-        long result = Math.multiplyExact(first.getValue(), second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"INT", "LARGEINT"}, returnType = "LARGEINT")
-    public static LargeIntLiteral multiplyIntLargeInt(IntegerLiteral first, LargeIntLiteral second) {
-        BigInteger result = second.getValue().multiply(new BigInteger(first.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"BIGINT", "TINYINT"}, returnType = "BIGINT")
-    public static BigIntLiteral multiplyBigIntTinyInt(BigIntLiteral first, TinyIntLiteral second) {
-        long result = Math.multiplyExact(first.getValue(), second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"BIGINT", "SMALLINT"}, returnType = "BIGINT")
-    public static BigIntLiteral multiplyBigIntSmallInt(BigIntLiteral first, SmallIntLiteral second) {
-        long result = Math.multiplyExact(first.getValue(), second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"BIGINT", "INT"}, returnType = "BIGINT")
-    public static BigIntLiteral multiplyBigIntInt(BigIntLiteral first, IntegerLiteral second) {
-        long result = Math.multiplyExact(first.getValue(), second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"BIGINT", "BIGINT"}, returnType = "BIGINT")
-    public static BigIntLiteral multiplyBigIntBigInt(BigIntLiteral first, BigIntLiteral second) {
-        long result = Math.multiplyExact(first.getValue(), second.getValue());
-        return new BigIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"BIGINT", "LARGEINT"}, returnType = "LARGEINT")
-    public static LargeIntLiteral multiplyBigIntLargeInt(BigIntLiteral first, LargeIntLiteral second) {
-        BigInteger result = second.getValue().multiply(new BigInteger(first.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"LARGEINT", "TINYINT"}, returnType = "BIGINT")
-    public static LargeIntLiteral multiplyLargeIntTinyInt(LargeIntLiteral first, TinyIntLiteral second) {
-        BigInteger result = first.getValue().multiply(new BigInteger(second.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"LARGEINT", "SMALLINT"}, returnType = "BIGINT")
-    public static LargeIntLiteral multiplyLargeIntSmallInt(LargeIntLiteral first, SmallIntLiteral second) {
-        BigInteger result = first.getValue().multiply(new BigInteger(second.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"LARGEINT", "INT"}, returnType = "BIGINT")
-    public static LargeIntLiteral multiplyLargeIntInt(LargeIntLiteral first, IntegerLiteral second) {
-        BigInteger result = first.getValue().multiply(new BigInteger(second.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"LARGEINT", "BIGINT"}, returnType = "BIGINT")
-    public static LargeIntLiteral multiplyLargeIntBigInt(LargeIntLiteral first, BigIntLiteral second) {
-        BigInteger result = first.getValue().multiply(new BigInteger(second.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"LARGEINT", "LARGEINT"}, returnType = "LARGEINT")
-    public static LargeIntLiteral multiplyLargeIntLargeInt(LargeIntLiteral first, LargeIntLiteral second) {
-        BigInteger result = first.getValue().multiply(new BigInteger(second.getValue().toString()));
-        return new LargeIntLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"DOUBLE", "DOUBLE"}, returnType = "DOUBLE")
-    public static DoubleLiteral multiplyDoubleDouble(DoubleLiteral first, DoubleLiteral second) {
-        double result = first.getValue() * second.getValue();
-        return new DoubleLiteral(result);
-    }
-
-    @ExecFunction(name = "multiply", argTypes = {"DECIMAL", "DECIMAL"}, returnType = "DECIMAL")
-    public static DecimalLiteral multiplyDecimalDecimal(DecimalLiteral first, DecimalLiteral second) {
-        BigDecimal result = first.getValue().multiply(second.getValue());
-        return new DecimalLiteral(result);
-    }
-
-    /**
-     * decimalV3 multiply in FE
-     */
-    @ExecFunction(name = "multiply", argTypes = {"DECIMALV3", "DECIMALV3"}, returnType = "DECIMALV3")
-    public static DecimalV3Literal multiplyDecimalV3DecimalV3(DecimalV3Literal first, DecimalV3Literal second) {
-        BigDecimal result = first.getValue().multiply(second.getValue());
-        DecimalV3Type t1 = (DecimalV3Type) first.getDataType();
-        DecimalV3Type t2 = (DecimalV3Type) second.getDataType();
-        int precision = t1.getPrecision() + t2.getPrecision();
-        int scale = t1.getScale() + t2.getScale();
-        return new DecimalV3Literal(DecimalV3Type.createDecimalV3Type(precision, scale), result);
-    }
-
-    /**
-     * Executable arithmetic functions divide
-     */
-    @ExecFunction(name = "divide", argTypes = {"DOUBLE", "DOUBLE"}, returnType = "DOUBLE")
-    public static Literal divideDouble(DoubleLiteral first, DoubleLiteral second) {
-        if (second.getValue() == 0.0) {
-            return new NullLiteral(first.getDataType());
-        }
-        double result = first.getValue() / second.getValue();
-        return new DoubleLiteral(result);
-    }
-
-    @ExecFunction(name = "divide", argTypes = {"DECIMAL", "DECIMAL"}, returnType = "DECIMAL")
-    public static Literal divideDecimal(DecimalLiteral first, DecimalLiteral second) {
-        if (first.getValue().compareTo(BigDecimal.ZERO) == 0) {
-            return new NullLiteral(first.getDataType());
-        }
-        BigDecimal result = first.getValue().divide(second.getValue());
-        return new DecimalLiteral(result);
-    }
-
-    /**
-     * decimalv3 divide in FE
-     */
-    @ExecFunction(name = "divide", argTypes = {"DECIMALV3", "DECIMALV3"}, returnType = "DECIMALV3")
-    public static Literal divideDecimalV3(DecimalV3Literal first, DecimalV3Literal second) {
-        if (first.getValue().compareTo(BigDecimal.ZERO) == 0) {
-            return new NullLiteral(first.getDataType());
-        }
-        DecimalV3Type t1 = (DecimalV3Type) first.getDataType();
-        DecimalV3Type t2 = (DecimalV3Type) second.getDataType();
-        BigDecimal result = first.getValue().divide(second.getValue());
-        return new DecimalV3Literal(DecimalV3Type.createDecimalV3Type(
-                t1.getPrecision(), t1.getScale() - t2.getScale()), result);
-    }
-
-    /**
-     * datetime arithmetic function date-add.
-     */
-    @ExecFunction(name = "date_add", argTypes = {"DATE", "INT"}, returnType = "DATE")
-    public static DateLiteral dateAdd(DateLiteral date, IntegerLiteral day) throws AnalysisException {
-        return daysAdd(date, day);
-    }
-
-    @ExecFunction(name = "date_add", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral dateAdd(DateTimeLiteral date, IntegerLiteral day) throws AnalysisException {
-        return daysAdd(date, day);
-    }
-
-    @ExecFunction(name = "date_add", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal dateAdd(DateV2Literal date, IntegerLiteral day) throws AnalysisException {
-        return daysAdd(date, day);
-    }
-
-    @ExecFunction(name = "date_add", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal dateAdd(DateTimeV2Literal date, IntegerLiteral day) throws AnalysisException {
-        return daysAdd(date, day);
-    }
-
-    /**
-     * datetime arithmetic function date-sub.
-     */
-    @ExecFunction(name = "date_sub", argTypes = {"DATE", "INT"}, returnType = "DATE")
-    public static DateLiteral dateSub(DateLiteral date, IntegerLiteral day) throws AnalysisException {
-        return dateAdd(date, new IntegerLiteral(-day.getValue()));
-    }
-
-    @ExecFunction(name = "date_sub", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral dateSub(DateTimeLiteral date, IntegerLiteral day) throws AnalysisException {
-        return dateAdd(date, new IntegerLiteral(-day.getValue()));
-    }
-
-    @ExecFunction(name = "date_sub", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal dateSub(DateV2Literal date, IntegerLiteral day) throws AnalysisException {
-        return dateAdd(date, new IntegerLiteral(-day.getValue()));
-    }
-
-    @ExecFunction(name = "date_sub", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal dateSub(DateTimeV2Literal date, IntegerLiteral day) throws AnalysisException {
-        return dateAdd(date, new IntegerLiteral(-day.getValue()));
-    }
-
-    /**
-     * datetime arithmetic function years-add.
-     */
-    @ExecFunction(name = "years_add", argTypes = {"DATE", "INT"}, returnType = "DATE")
-    public static DateLiteral yearsAdd(DateLiteral date, IntegerLiteral year) throws AnalysisException {
-        return date.plusYears(year.getValue());
-    }
-
-    @ExecFunction(name = "years_add", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral yearsAdd(DateTimeLiteral date, IntegerLiteral year) throws AnalysisException {
-        return date.plusYears(year.getValue());
-    }
-
-    @ExecFunction(name = "years_add", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal yearsAdd(DateV2Literal date, IntegerLiteral year) throws AnalysisException {
-        return date.plusYears(year.getValue());
-    }
-
-    @ExecFunction(name = "years_add", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal yearsAdd(DateTimeV2Literal date, IntegerLiteral year) throws AnalysisException {
-        return date.plusYears(year.getValue());
-    }
-
-    /**
-     * datetime arithmetic function months-add.
-     */
-    @ExecFunction(name = "months_add", argTypes = {"DATE", "INT"}, returnType = "DATE")
-    public static DateLiteral monthsAdd(DateLiteral date, IntegerLiteral month) throws AnalysisException {
-        return date.plusMonths(month.getValue());
-    }
-
-    @ExecFunction(name = "months_add", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral monthsAdd(DateTimeLiteral date, IntegerLiteral month) throws AnalysisException {
-        return date.plusMonths(month.getValue());
-    }
-
-    @ExecFunction(name = "months_add", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal monthsAdd(DateV2Literal date, IntegerLiteral month) throws AnalysisException {
-        return date.plusMonths(month.getValue());
-    }
-
-    @ExecFunction(name = "months_add", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal monthsAdd(DateTimeV2Literal date, IntegerLiteral month) throws AnalysisException {
-        return date.plusMonths(month.getValue());
-    }
-
-    /**
-     * datetime arithmetic function days-add.
-     */
-    @ExecFunction(name = "days_add", argTypes = {"DATE", "INT"}, returnType = "DATE")
-    public static DateLiteral daysAdd(DateLiteral date, IntegerLiteral day) throws AnalysisException {
-        return date.plusDays(day.getValue());
-    }
-
-    @ExecFunction(name = "days_add", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral daysAdd(DateTimeLiteral date, IntegerLiteral day) throws AnalysisException {
-        return date.plusDays(day.getValue());
-    }
-
-    @ExecFunction(name = "days_add", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal daysAdd(DateV2Literal date, IntegerLiteral day) throws AnalysisException {
-        return date.plusDays(day.getValue());
-    }
-
-    @ExecFunction(name = "days_add", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal daysAdd(DateTimeV2Literal date, IntegerLiteral day) throws AnalysisException {
-        return date.plusDays(day.getValue());
-    }
-
-    /**
-     * datetime arithmetic function hours-add.
-     */
-    @ExecFunction(name = "hours_add", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral hoursAdd(DateTimeLiteral date, IntegerLiteral hour) throws AnalysisException {
-        return date.plusHours(hour.getValue());
-    }
-
-    @ExecFunction(name = "hours_add", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal hoursAdd(DateTimeV2Literal date, IntegerLiteral hour) throws AnalysisException {
-        return date.plusHours(hour.getValue());
-    }
-
-    /**
-     * datetime arithmetic function minutes-add.
-     */
-    @ExecFunction(name = "minutes_add", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral minutesAdd(DateTimeLiteral date, IntegerLiteral minute) throws AnalysisException {
-        return date.plusMinutes(minute.getValue());
-    }
-
-    @ExecFunction(name = "minutes_add", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal minutesAdd(DateTimeV2Literal date, IntegerLiteral minute) throws AnalysisException {
-        return date.plusMinutes(minute.getValue());
-    }
-
-    /**
-     * datetime arithmetic function seconds-add.
-     */
-    @ExecFunction(name = "seconds_add", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral secondsAdd(DateTimeLiteral date, IntegerLiteral second) throws AnalysisException {
-        return date.plusSeconds(second.getValue());
-    }
-
-    @ExecFunction(name = "seconds_add", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal secondsAdd(DateTimeV2Literal date, IntegerLiteral second) throws AnalysisException {
-        return date.plusSeconds(second.getValue());
-    }
-
-    /**
-     * datetime arithmetic function years-sub.
-     */
-    @ExecFunction(name = "years_sub", argTypes = {"DATE", "INT"}, returnType = "DATE")
-    public static DateLiteral yearsSub(DateLiteral date, IntegerLiteral year) throws AnalysisException {
-        return yearsAdd(date, new IntegerLiteral(-year.getValue()));
-    }
-
-    @ExecFunction(name = "years_sub", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral yearsSub(DateTimeLiteral date, IntegerLiteral year) throws AnalysisException {
-        return yearsAdd(date, new IntegerLiteral(-year.getValue()));
-    }
-
-    @ExecFunction(name = "years_sub", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal yearsSub(DateV2Literal date, IntegerLiteral year) throws AnalysisException {
-        return yearsAdd(date, new IntegerLiteral(-year.getValue()));
-    }
-
-    @ExecFunction(name = "years_sub", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal yearsSub(DateTimeV2Literal date, IntegerLiteral year) throws AnalysisException {
-        return yearsAdd(date, new IntegerLiteral(-year.getValue()));
-    }
-
-    /**
-     * datetime arithmetic function months-sub
-     */
-    @ExecFunction(name = "months_sub", argTypes = {"DATE", "INT"}, returnType = "DATE")
-    public static DateLiteral monthsSub(DateLiteral date, IntegerLiteral month) throws AnalysisException {
-        return monthsAdd(date, new IntegerLiteral(-month.getValue()));
-    }
-
-    @ExecFunction(name = "months_sub", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral monthsSub(DateTimeLiteral date, IntegerLiteral month) throws AnalysisException {
-        return monthsAdd(date, new IntegerLiteral(-month.getValue()));
-    }
-
-    @ExecFunction(name = "months_sub", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal monthsSub(DateV2Literal date, IntegerLiteral month) throws AnalysisException {
-        return monthsAdd(date, new IntegerLiteral(-month.getValue()));
-    }
-
-    @ExecFunction(name = "months_sub", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal monthsSub(DateTimeV2Literal date, IntegerLiteral month) throws AnalysisException {
-        return monthsAdd(date, new IntegerLiteral(-month.getValue()));
-    }
-
-    /**
-     * datetime arithmetic function days-sub
-     */
-    @ExecFunction(name = "days_sub", argTypes = {"DATE", "INT"}, returnType = "DATE")
-    public static DateLiteral daysSub(DateLiteral date, IntegerLiteral day) throws AnalysisException {
-        return daysAdd(date, new IntegerLiteral(-day.getValue()));
-    }
-
-    @ExecFunction(name = "days_sub", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral daysSub(DateTimeLiteral date, IntegerLiteral day) throws AnalysisException {
-        return daysAdd(date, new IntegerLiteral(-day.getValue()));
-    }
-
-    @ExecFunction(name = "days_sub", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal daysSub(DateV2Literal date, IntegerLiteral day) throws AnalysisException {
-        return daysAdd(date, new IntegerLiteral(-day.getValue()));
-    }
-
-    @ExecFunction(name = "days_sub", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal daysSub(DateTimeV2Literal date, IntegerLiteral day) throws AnalysisException {
-        return daysAdd(date, new IntegerLiteral(-day.getValue()));
-    }
-
-    /**
-     * datetime arithmetic function hours-sub
-     */
-    @ExecFunction(name = "hours_sub", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral hoursSub(DateTimeLiteral date, IntegerLiteral hour) throws AnalysisException {
-        return hoursAdd(date, new IntegerLiteral(-hour.getValue()));
-    }
-
-    @ExecFunction(name = "hours_sub", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal hoursSub(DateTimeV2Literal date, IntegerLiteral hour) throws AnalysisException {
-        return hoursAdd(date, new IntegerLiteral(-hour.getValue()));
-    }
-
-    /**
-     * datetime arithmetic function minutes-sub
-     */
-    @ExecFunction(name = "minutes_sub", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral minutesSub(DateTimeLiteral date, IntegerLiteral minute) throws AnalysisException {
-        return minutesAdd(date, new IntegerLiteral(-minute.getValue()));
-    }
-
-    @ExecFunction(name = "minutes_sub", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal minutesSub(DateTimeV2Literal date, IntegerLiteral minute) throws AnalysisException {
-        return minutesAdd(date, new IntegerLiteral(-minute.getValue()));
-    }
-
-    /**
-     * datetime arithmetic function seconds-sub
-     */
-    @ExecFunction(name = "seconds_sub", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral secondsSub(DateTimeLiteral date, IntegerLiteral second) throws AnalysisException {
-        return secondsAdd(date, new IntegerLiteral(-second.getValue()));
-    }
-
-    @ExecFunction(name = "seconds_sub", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal secondsSub(DateTimeV2Literal date, IntegerLiteral second) throws AnalysisException {
-        return secondsAdd(date, new IntegerLiteral(-second.getValue()));
-    }
-
-    /**
-     * datetime arithmetic function date
-     */
-    @ExecFunction(name = "date", argTypes = {"DATETIME"}, returnType = "DATE")
-    public static DateLiteral date(DateTimeLiteral dateTime) throws AnalysisException {
-        return new DateLiteral(dateTime.getYear(), dateTime.getMonth(), dateTime.getDay());
-    }
-
-    @ExecFunction(name = "date", argTypes = {"DATETIMEV2"}, returnType = "DATE")
-    public static DateLiteral date(DateTimeV2Literal dateTime) throws AnalysisException {
-        return new DateLiteral(dateTime.getYear(), dateTime.getMonth(), dateTime.getDay());
-    }
-
-    /**
-     * datetime arithmetic function datediff
-     */
-    @ExecFunction(name = "datediff", argTypes = {"DATETIMEV2", "DATETIMEV2"}, returnType = "INT")
-    public static IntegerLiteral dateDiff(DateTimeV2Literal date1, DateTimeV2Literal date2) {
-        return new IntegerLiteral(dateDiff(date1.getYear(), date1.getMonth(), date1.getDay(),
-                date2.getYear(), date2.getMonth(), date2.getDay()));
-    }
-
-    @ExecFunction(name = "datediff", argTypes = {"DATEV2", "DATEV2"}, returnType = "INT")
-    public static IntegerLiteral dateDiff(DateV2Literal date1, DateV2Literal date2) {
-        return new IntegerLiteral(dateDiff(date1.getYear(), date1.getMonth(), date1.getDay(),
-                date2.getYear(), date2.getMonth(), date2.getDay()));
-    }
-
-    @ExecFunction(name = "datediff", argTypes = {"DATETIMEV2", "DATEV2"}, returnType = "INT")
-    public static IntegerLiteral dateDiff(DateTimeV2Literal date1, DateV2Literal date2) {
-        return new IntegerLiteral(dateDiff(date1.getYear(), date1.getMonth(), date1.getDay(),
-                date2.getYear(), date2.getMonth(), date2.getDay()));
-    }
-
-    @ExecFunction(name = "datediff", argTypes = {"DATEV2", "DATETIMEV2"}, returnType = "INT")
-    public static IntegerLiteral dateDiff(DateV2Literal date1, DateTimeV2Literal date2) {
-        return new IntegerLiteral(dateDiff(date1.getYear(), date1.getMonth(), date1.getDay(),
-                date2.getYear(), date2.getMonth(), date2.getDay()));
-    }
-
-    @ExecFunction(name = "datediff", argTypes = {"DATETIME", "DATETIME"}, returnType = "INT")
-    public static IntegerLiteral dateDiff(DateTimeLiteral date1, DateTimeLiteral date2) {
-        return new IntegerLiteral(dateDiff(date1.getYear(), date1.getMonth(), date1.getDay(),
-                date2.getYear(), date2.getMonth(), date2.getDay()));
-    }
-
-    private static int dateDiff(long year1, long month1, long day1, long year2, long month2, long day2) {
-        Calendar calendar = Calendar.getInstance();
-        calendar.set(((int) year1), ((int) month1), ((int) day1));
-        long time1 = calendar.getTimeInMillis();
-        calendar.set(((int) year2), ((int) month2), ((int) day2));
-        long time2 = calendar.getTimeInMillis();
-        return ((int) ((time1 - time2) / (1000 * 3600 * 24)));
-    }
-
-    /**
-     * datetime arithmetic function date-format
-     */
-    @ExecFunction(name = "date_format", argTypes = {"DATE", "VARCHAR"}, returnType = "VARCHAR")
-    public static VarcharLiteral dateFormat(DateLiteral date, VarcharLiteral format) {
-        return new VarcharLiteral(DateUtils.formatBuilder(format.getValue()).toFormatter().format(
-                java.time.LocalDate.of(((int) date.getYear()), ((int) date.getMonth()), ((int) date.getDay()))));
-    }
-
-    @ExecFunction(name = "date_format", argTypes = {"DATETIME", "VARCHAR"}, returnType = "VARCHAR")
-    public static VarcharLiteral dateFormat(DateTimeLiteral date, VarcharLiteral format) {
-        return new VarcharLiteral(DateUtils.formatBuilder(format.getValue()).toFormatter().format(
-                java.time.LocalDateTime.of(((int) date.getYear()), ((int) date.getMonth()), ((int) date.getDay()),
-                        ((int) date.getHour()), ((int) date.getMinute()), ((int) date.getSecond()))));
-    }
-
-    @ExecFunction(name = "date_format", argTypes = {"DATEV2", "VARCHAR"}, returnType = "VARCHAR")
-    public static VarcharLiteral dateFormat(DateV2Literal date, VarcharLiteral format) {
-        return new VarcharLiteral(DateUtils.formatBuilder(format.getValue()).toFormatter().format(
-                java.time.LocalDate.of(((int) date.getYear()), ((int) date.getMonth()), ((int) date.getDay()))));
-    }
-
-    @ExecFunction(name = "date_format", argTypes = {"DATETIMEV2", "VARCHAR"}, returnType = "VARCHAR")
-    public static VarcharLiteral dateFormat(DateTimeV2Literal date, VarcharLiteral format) {
-        return new VarcharLiteral(DateUtils.formatBuilder(format.getValue()).toFormatter().format(
-                java.time.LocalDateTime.of(((int) date.getYear()), ((int) date.getMonth()), ((int) date.getDay()),
-                        ((int) date.getHour()), ((int) date.getMinute()), ((int) date.getSecond()))));
-    }
-
-    /**
-     * datetime arithmetic function day-of-week
-     */
-    @ExecFunction(name = "dayofweek", argTypes = {"DATE"}, returnType = "INT")
-    public static IntegerLiteral dayOfWeek(DateLiteral date) {
-        return new IntegerLiteral(distanceToFirstDayOfWeek(date.getYear(), date.getMonth(), date.getDay()));
-    }
-
-    @ExecFunction(name = "dayofweek", argTypes = {"DATETIME"}, returnType = "INT")
-    public static IntegerLiteral dayOfWeek(DateTimeLiteral date) {
-        return new IntegerLiteral(distanceToFirstDayOfWeek(date.getYear(), date.getMonth(), date.getDay()));
-    }
-
-    @ExecFunction(name = "dayofweek", argTypes = {"DATEV2"}, returnType = "INT")
-    public static IntegerLiteral dayOfWeek(DateV2Literal date) {
-        return new IntegerLiteral(distanceToFirstDayOfWeek(date.getYear(), date.getMonth(), date.getDay()));
-    }
-
-    @ExecFunction(name = "dayofweek", argTypes = {"DATETIMEV2"}, returnType = "INT")
-    public static IntegerLiteral dayOfWeek(DateTimeV2Literal date) {
-        return new IntegerLiteral(distanceToFirstDayOfWeek(date.getYear(), date.getMonth(), date.getDay()));
-    }
-
-    private static int distanceToFirstDayOfWeek(long year, long month, long day) {
-        return LocalDate.of((int) year, (int) month, (int) day).getDayOfWeek().getValue() - 1;
-    }
-
-    private static long[] firstDayOfWeek(long year, long month, long day) {
-        int distance = distanceToFirstDayOfWeek(year, month, day);
-        DateLiteral temp = new DateLiteral(year, month, day).plusDays(-distance);
-        return new long[] {temp.getYear(), temp.getMonth(), temp.getDay()};
-    }
-
-    /**
-     * datetime arithmetic function date-trunc
-     */
-    @ExecFunction(name = "date_trunc", argTypes = {"DATETIME", "VARCHAR"}, returnType = "DATETIME")
-    public static DateTimeLiteral dateTrunc(DateTimeLiteral date, VarcharLiteral trunc) {
-        long[] timeTags = dateTruncHelper(date.getYear(), date.getMonth(), date.getDay(),
-                date.getHour(), date.getMinute(), date.getSecond(), trunc.getValue());
-        return new DateTimeLiteral(timeTags[0], timeTags[1], timeTags[2], timeTags[3], timeTags[4], timeTags[5]);
-    }
-
-    @ExecFunction(name = "date_trunc", argTypes = {"DATETIMEV2", "VARCHAR"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal dateTrunc(DateTimeV2Literal date, VarcharLiteral trunc) {
-        long[] timeTags = dateTruncHelper(date.getYear(), date.getMonth(), date.getDay(),
-                date.getHour(), date.getMinute(), date.getSecond(), trunc.getValue());
-        return new DateTimeV2Literal(timeTags[0], timeTags[1], timeTags[2], timeTags[3], timeTags[4], timeTags[5], 0);
-    }
-
-    private static long[] dateTruncHelper(long year, long month, long day, long hour, long minute, long second,
-            String trunc) {
-        switch (trunc.toLowerCase()) {
-            case "year":
-                month = 0;
-            case "quarter": // CHECKSTYLE IGNORE THIS LINE
-                month = ((month - 1) / 4) * 4 + 1;
-            case "month": // CHECKSTYLE IGNORE THIS LINE
-                day = 1;
-                break;
-            case "week":
-                long[] ymd = firstDayOfWeek(year, month, day);
-                year = ymd[0];
-                month = ymd[1];
-                day = ymd[2];
-            default: // CHECKSTYLE IGNORE THIS LINE
-                break;
-        }
-        switch (trunc.toLowerCase()) {
-            default: // CHECKSTYLE IGNORE THIS LINE
-            case "day": // CHECKSTYLE IGNORE THIS LINE
-                hour = 0;
-            case "hour": // CHECKSTYLE IGNORE THIS LINE
-                minute = 0;
-            case "minute": // CHECKSTYLE IGNORE THIS LINE
-                second = 0;
-        }
-        return new long[] {year, month, day, hour, minute, second};
-    }
-
-    /**
-     * datetime arithmetic function date-v2
-     */
-    @ExecFunction(name = "datev2", argTypes = {"DATETIMEV2"}, returnType = "DATEV2")
-    public static DateV2Literal dateV2(DateTimeV2Literal dateTime) {
-        return new DateV2Literal(dateTime.getYear(), dateTime.getMonth(), dateTime.getDay());
-    }
-
-    /**
-     * datetime arithmetic function date-ceil and date-floor series helper function
-     */
-    private static LocalDateTime getDateFloor(DATE tag, LocalDateTime date, int period, LocalDateTime origin,
-            boolean getCeil) {
-        // Algorithm:
-        // Firstly, get the unit distance of the two date.
-        // Secondly, if the origin date is bigger than the date, subtract it to a date before the date by unit.
-        // Thirdly, re-calculate the distance of the two date.
-        // Fourthly, get the ceil and floor date of the date by unit and select the corresponding date as the answer.
-
-        // handle origin > date
-        if (origin.compareTo(date) > 0) {
-            Duration duration = Duration.between(date, origin);
-            long hour = Math.abs(duration.toHours());
-            long ceil = ((hour - 1) / period + 1) * period;
-            origin = origin.minusHours(ceil);
-        }
-
-        // get distance
-        Duration duration = Duration.between(origin, date);
-        long hour = Math.abs(duration.toHours());
-        long ceil = ((hour - 1) / period + 1) * period;
-        long floor = hour / period * period;
-        LocalDateTime floorDate = origin.plusHours(floor);
-        LocalDateTime ceilDate = origin.plusHours(ceil);
-
-        return getCeil ? ceilDate : floorDate;
-    }
-
-    /**
-     * datetime arithmetic function year-ceil
-     */
-    @ExecFunction(name = "year_ceil", argTypes = {"DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral yearCeil(DateTimeLiteral date) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), 1, START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "year_ceil", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral yearCeil(DateTimeLiteral date, IntegerLiteral period) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "year_ceil", argTypes = {"DATETIME", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral yearCeil(DateTimeLiteral date, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), 1, origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "year_ceil", argTypes = {"DATETIME", "INT", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral yearCeil(DateTimeLiteral date, IntegerLiteral period, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "year_ceil", argTypes = {"DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal yearCeil(DateV2Literal date) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), 1, START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "year_ceil", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal yearCeil(DateV2Literal date, IntegerLiteral period) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "year_ceil", argTypes = {"DATEV2", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal yearCeil(DateV2Literal date, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), 1, origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "year_ceil", argTypes = {"DATEV2", "INT", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal yearCeil(DateV2Literal date, IntegerLiteral period, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "year_ceil", argTypes = {"DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal yearCeil(DateTimeV2Literal date) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), 1, START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "year_ceil", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal yearCeil(DateTimeV2Literal date, IntegerLiteral period) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "year_ceil", argTypes = {"DATETIMEV2", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal yearCeil(DateTimeV2Literal date, DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), 1, origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "year_ceil", argTypes = {"DATETIMEV2", "INT", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal yearCeil(DateTimeV2Literal date, IntegerLiteral period, DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), true));
-    }
-
-    /**
-     * datetime arithmetic function month-ceil
-     */
-    @ExecFunction(name = "month_ceil", argTypes = {"DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral monthCeil(DateTimeLiteral date) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), 1, START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "month_ceil", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral monthCeil(DateTimeLiteral date, IntegerLiteral period) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "month_ceil", argTypes = {"DATETIME", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral monthCeil(DateTimeLiteral date, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), 1, origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "month_ceil", argTypes = {"DATETIME", "INT", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral monthCeil(DateTimeLiteral date, IntegerLiteral period, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "month_ceil", argTypes = {"DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal monthCeil(DateV2Literal date) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), 1, START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "month_ceil", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal monthCeil(DateV2Literal date, IntegerLiteral period) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "month_ceil", argTypes = {"DATEV2", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal monthCeil(DateV2Literal date, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), 1, origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "month_ceil", argTypes = {"DATEV2", "INT", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal monthCeil(DateV2Literal date, IntegerLiteral period, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "month_ceil", argTypes = {"DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal monthCeil(DateTimeV2Literal date) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), 1, START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "month_ceil", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal monthCeil(DateTimeV2Literal date, IntegerLiteral period) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "month_ceil", argTypes = {"DATETIMEV2", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal monthCeil(DateTimeV2Literal date, DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), 1, origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "month_ceil", argTypes = {"DATETIMEV2", "INT", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal monthCeil(DateTimeV2Literal date, IntegerLiteral period, DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), true));
-    }
-
-    /**
-     * datetime arithmetic function day-ceil
-     */
-    @ExecFunction(name = "day_ceil", argTypes = {"DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral dayCeil(DateTimeLiteral date) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), 1, START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "day_ceil", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral dayCeil(DateTimeLiteral date, IntegerLiteral period) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "day_ceil", argTypes = {"DATETIME", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral dayCeil(DateTimeLiteral date, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), 1, origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "day_ceil", argTypes = {"DATETIME", "INT", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral dayCeil(DateTimeLiteral date, IntegerLiteral period, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "day_ceil", argTypes = {"DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal dayCeil(DateV2Literal date) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), 1, START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "day_ceil", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal dayCeil(DateV2Literal date, IntegerLiteral period) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "day_ceil", argTypes = {"DATEV2", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal dayCeil(DateV2Literal date, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), 1, origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "day_ceil", argTypes = {"DATEV2", "INT", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal dayCeil(DateV2Literal date, IntegerLiteral period, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "day_ceil", argTypes = {"DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal dayCeil(DateTimeV2Literal date) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), 1, START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "day_ceil", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal dayCeil(DateTimeV2Literal date, IntegerLiteral period) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "day_ceil", argTypes = {"DATETIMEV2", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal dayCeil(DateTimeV2Literal date, DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), 1, origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "day_ceil", argTypes = {"DATETIMEV2", "INT", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal dayCeil(DateTimeV2Literal date, IntegerLiteral period, DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), true));
-    }
-
-    /**
-     * datetime arithmetic function hour-ceil
-     */
-    @ExecFunction(name = "hour_ceil", argTypes = {"DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral hourCeil(DateTimeLiteral date) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), 1, START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "hour_ceil", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral hourCeil(DateTimeLiteral date, IntegerLiteral period) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "hour_ceil", argTypes = {"DATETIME", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral hourCeil(DateTimeLiteral date, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), 1, origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "hour_ceil", argTypes = {"DATETIME", "INT", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral hourCeil(DateTimeLiteral date, IntegerLiteral period, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "hour_ceil", argTypes = {"DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal hourCeil(DateV2Literal date) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), 1, START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "hour_ceil", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal hourCeil(DateV2Literal date, IntegerLiteral period) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "hour_ceil", argTypes = {"DATEV2", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal hourCeil(DateV2Literal date, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), 1, origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "hour_ceil", argTypes = {"DATEV2", "INT", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal hourCeil(DateV2Literal date, IntegerLiteral period, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "hour_ceil", argTypes = {"DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal hourCeil(DateTimeV2Literal date) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), 1, START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "hour_ceil", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal hourCeil(DateTimeV2Literal date, IntegerLiteral period) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "hour_ceil", argTypes = {"DATETIMEV2", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal hourCeil(DateTimeV2Literal date, DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), 1, origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "hour_ceil", argTypes = {"DATETIMEV2", "INT", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal hourCeil(DateTimeV2Literal date, IntegerLiteral period, DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), true));
-    }
-
-    /**
-     * datetime arithmetic function minute-ceil
-     */
-    @ExecFunction(name = "minute_ceil", argTypes = {"DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral minuteCeil(DateTimeLiteral date) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), 1, START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "minute_ceil", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral minuteCeil(DateTimeLiteral date, IntegerLiteral period) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "minute_ceil", argTypes = {"DATETIME", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral minuteCeil(DateTimeLiteral date, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), 1, origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "minute_ceil", argTypes = {"DATETIME", "INT", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral minuteCeil(DateTimeLiteral date, IntegerLiteral period, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "minute_ceil", argTypes = {"DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal minuteCeil(DateV2Literal date) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), 1, START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "minute_ceil", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal minuteCeil(DateV2Literal date, IntegerLiteral period) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "minute_ceil", argTypes = {"DATEV2", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal minuteCeil(DateV2Literal date, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), 1, origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "minute_ceil", argTypes = {"DATEV2", "INT", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal minuteCeil(DateV2Literal date, IntegerLiteral period, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "minute_ceil", argTypes = {"DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal minuteCeil(DateTimeV2Literal date) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), 1, START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "minute_ceil", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal minuteCeil(DateTimeV2Literal date, IntegerLiteral period) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "minute_ceil", argTypes = {"DATETIMEV2", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal minuteCeil(DateTimeV2Literal date, DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), 1, origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "minute_ceil", argTypes = {"DATETIMEV2", "INT", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal minuteCeil(DateTimeV2Literal date, IntegerLiteral period,
-            DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), true));
-    }
-
-    /**
-     * datetime arithmetic function SECOND-ceil
-     */
-    @ExecFunction(name = "second_ceil", argTypes = {"DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral secondCeil(DateTimeLiteral date) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), 1, START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "second_ceil", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral secondCeil(DateTimeLiteral date, IntegerLiteral period) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "second_ceil", argTypes = {"DATETIME", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral secondCeil(DateTimeLiteral date, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), 1, origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "second_ceil", argTypes = {"DATETIME", "INT", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral secondCeil(DateTimeLiteral date, IntegerLiteral period, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "second_ceil", argTypes = {"DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal secondCeil(DateV2Literal date) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), 1, START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "second_ceil", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal secondCeil(DateV2Literal date, IntegerLiteral period) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "second_ceil", argTypes = {"DATEV2", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal secondCeil(DateV2Literal date, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), 1, origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "second_ceil", argTypes = {"DATEV2", "INT", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal secondCeil(DateV2Literal date, IntegerLiteral period, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "second_ceil", argTypes = {"DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal secondCeil(DateTimeV2Literal date) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), 1, START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "second_ceil", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal secondCeil(DateTimeV2Literal date, IntegerLiteral period) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, true));
-    }
-
-    @ExecFunction(name = "second_ceil", argTypes = {"DATETIMEV2", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal secondCeil(DateTimeV2Literal date, DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), 1, origin.toJavaDateType(), true));
-    }
-
-    @ExecFunction(name = "second_ceil", argTypes = {"DATETIMEV2", "INT", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal secondCeil(DateTimeV2Literal date, IntegerLiteral period,
-            DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), true));
-    }
-
-    /**
-     * datetime arithmetic function year-floor
-     */
-    @ExecFunction(name = "year_floor", argTypes = {"DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral yearFloor(DateTimeLiteral date) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), 1, START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "year_floor", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral yearFloor(DateTimeLiteral date, IntegerLiteral period) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "year_floor", argTypes = {"DATETIME", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral yearFloor(DateTimeLiteral date, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), 1, origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "year_floor", argTypes = {"DATETIME", "INT", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral yearFloor(DateTimeLiteral date, IntegerLiteral period, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "year_floor", argTypes = {"DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal yearFloor(DateV2Literal date) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), 1, START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "year_floor", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal yearFloor(DateV2Literal date, IntegerLiteral period) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "year_floor", argTypes = {"DATEV2", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal yearFloor(DateV2Literal date, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), 1, origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "year_floor", argTypes = {"DATEV2", "INT", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal yearFloor(DateV2Literal date, IntegerLiteral period, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "year_floor", argTypes = {"DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal yearFloor(DateTimeV2Literal date) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), 1, START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "year_floor", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal yearFloor(DateTimeV2Literal date, IntegerLiteral period) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "year_floor", argTypes = {"DATETIMEV2", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal yearFloor(DateTimeV2Literal date, DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), 1, origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "year_floor", argTypes = {"DATETIMEV2", "INT", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal yearFloor(DateTimeV2Literal date, IntegerLiteral period, DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.YEAR, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), false));
-    }
-
-    /**
-     * datetime arithmetic function month-floor
-     */
-    @ExecFunction(name = "month_floor", argTypes = {"DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral monthFloor(DateTimeLiteral date) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), 1, START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "month_floor", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral monthFloor(DateTimeLiteral date, IntegerLiteral period) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "month_floor", argTypes = {"DATETIME", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral monthFloor(DateTimeLiteral date, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), 1, origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "month_floor", argTypes = {"DATETIME", "INT", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral monthFloor(DateTimeLiteral date, IntegerLiteral period, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "month_floor", argTypes = {"DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal monthFloor(DateV2Literal date) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), 1, START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "month_floor", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal monthFloor(DateV2Literal date, IntegerLiteral period) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "month_floor", argTypes = {"DATEV2", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal monthFloor(DateV2Literal date, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), 1, origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "month_floor", argTypes = {"DATEV2", "INT", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal monthFloor(DateV2Literal date, IntegerLiteral period, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "month_floor", argTypes = {"DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal monthFloor(DateTimeV2Literal date) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), 1, START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "month_floor", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal monthFloor(DateTimeV2Literal date, IntegerLiteral period) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "month_floor", argTypes = {"DATETIMEV2", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal monthFloor(DateTimeV2Literal date, DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), 1, origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "month_floor", argTypes = {"DATETIMEV2", "INT", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal monthFloor(DateTimeV2Literal date, IntegerLiteral period,
-            DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MONTH, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), false));
-    }
-
-    /**
-     * datetime arithmetic function day-floor
-     */
-    @ExecFunction(name = "day_floor", argTypes = {"DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral dayFloor(DateTimeLiteral date) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), 1, START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "day_floor", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral dayFloor(DateTimeLiteral date, IntegerLiteral period) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "day_floor", argTypes = {"DATETIME", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral dayFloor(DateTimeLiteral date, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), 1, origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "day_floor", argTypes = {"DATETIME", "INT", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral dayFloor(DateTimeLiteral date, IntegerLiteral period, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "day_floor", argTypes = {"DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal dayFloor(DateV2Literal date) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), 1, START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "day_floor", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal dayFloor(DateV2Literal date, IntegerLiteral period) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "day_floor", argTypes = {"DATEV2", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal dayFloor(DateV2Literal date, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), 1, origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "day_floor", argTypes = {"DATEV2", "INT", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal dayFloor(DateV2Literal date, IntegerLiteral period, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "day_floor", argTypes = {"DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal dayFloor(DateTimeV2Literal date) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), 1, START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "day_floor", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal dayFloor(DateTimeV2Literal date, IntegerLiteral period) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "day_floor", argTypes = {"DATETIMEV2", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal dayFloor(DateTimeV2Literal date, DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), 1, origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "day_floor", argTypes = {"DATETIMEV2", "INT", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal dayFloor(DateTimeV2Literal date, IntegerLiteral period, DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.DAY, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), false));
-    }
-
-    /**
-     * datetime arithmetic function hour-floor
-     */
-    @ExecFunction(name = "hour_floor", argTypes = {"DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral hourFloor(DateTimeLiteral date) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), 1, START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "hour_floor", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral hourFloor(DateTimeLiteral date, IntegerLiteral period) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "hour_floor", argTypes = {"DATETIME", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral hourFloor(DateTimeLiteral date, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), 1, origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "hour_floor", argTypes = {"DATETIME", "INT", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral hourFloor(DateTimeLiteral date, IntegerLiteral period, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "hour_floor", argTypes = {"DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal hourFloor(DateV2Literal date) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), 1, START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "hour_floor", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal hourFloor(DateV2Literal date, IntegerLiteral period) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "hour_floor", argTypes = {"DATEV2", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal hourFloor(DateV2Literal date, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), 1, origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "hour_floor", argTypes = {"DATEV2", "INT", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal hourFloor(DateV2Literal date, IntegerLiteral period, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "hour_floor", argTypes = {"DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal hourFloor(DateTimeV2Literal date) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), 1, START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "hour_floor", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal hourFloor(DateTimeV2Literal date, IntegerLiteral period) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "hour_floor", argTypes = {"DATETIMEV2", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal hourFloor(DateTimeV2Literal date, DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), 1, origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "hour_floor", argTypes = {"DATETIMEV2", "INT", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal hourFloor(DateTimeV2Literal date, IntegerLiteral period, DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.HOUR, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), false));
-    }
-
-    /**
-     * datetime arithmetic function minute-floor
-     */
-    @ExecFunction(name = "minute_floor", argTypes = {"DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral minuteFloor(DateTimeLiteral date) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), 1, START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "minute_floor", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral minuteFloor(DateTimeLiteral date, IntegerLiteral period) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "minute_floor", argTypes = {"DATETIME", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral minuteFloor(DateTimeLiteral date, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), 1, origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "minute_floor", argTypes = {"DATETIME", "INT", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral minuteFloor(DateTimeLiteral date, IntegerLiteral period, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "minute_floor", argTypes = {"DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal minuteFloor(DateV2Literal date) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), 1, START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "minute_floor", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal minuteFloor(DateV2Literal date, IntegerLiteral period) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "minute_floor", argTypes = {"DATEV2", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal minuteFloor(DateV2Literal date, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), 1, origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "minute_floor", argTypes = {"DATEV2", "INT", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal minuteFloor(DateV2Literal date, IntegerLiteral period, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "minute_floor", argTypes = {"DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal minuteFloor(DateTimeV2Literal date) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), 1, START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "minute_floor", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal minuteFloor(DateTimeV2Literal date, IntegerLiteral period) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "minute_floor", argTypes = {"DATETIMEV2", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal minuteFloor(DateTimeV2Literal date, DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), 1, origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "minute_floor", argTypes = {"DATETIMEV2", "INT", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal minuteFloor(DateTimeV2Literal date, IntegerLiteral period,
-            DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.MINUTE, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), false));
-    }
-
-    /**
-     * datetime arithmetic function SECOND-floor
-     */
-    @ExecFunction(name = "second_floor", argTypes = {"DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral secondFloor(DateTimeLiteral date) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), 1, START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "second_floor", argTypes = {"DATETIME", "INT"}, returnType = "DATETIME")
-    public static DateTimeLiteral secondFloor(DateTimeLiteral date, IntegerLiteral period) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "second_floor", argTypes = {"DATETIME", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral secondFloor(DateTimeLiteral date, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), 1, origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "second_floor", argTypes = {"DATETIME", "INT", "DATETIME"}, returnType = "DATETIME")
-    public static DateTimeLiteral secondFloor(DateTimeLiteral date, IntegerLiteral period, DateTimeLiteral origin) {
-        return DateTimeLiteral.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "second_floor", argTypes = {"DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal secondFloor(DateV2Literal date) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), 1, START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "second_floor", argTypes = {"DATEV2", "INT"}, returnType = "DATEV2")
-    public static DateV2Literal secondFloor(DateV2Literal date, IntegerLiteral period) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "second_floor", argTypes = {"DATEV2", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal secondFloor(DateV2Literal date, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), 1, origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "second_floor", argTypes = {"DATEV2", "INT", "DATEV2"}, returnType = "DATEV2")
-    public static DateV2Literal secondFloor(DateV2Literal date, IntegerLiteral period, DateV2Literal origin) {
-        return DateV2Literal.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "second_floor", argTypes = {"DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal secondFloor(DateTimeV2Literal date) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), 1, START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "second_floor", argTypes = {"DATETIMEV2", "INT"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal secondFloor(DateTimeV2Literal date, IntegerLiteral period) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), period.getValue(), START_ORIGINAL_DAY, false));
-    }
-
-    @ExecFunction(name = "second_floor", argTypes = {"DATETIMEV2", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal secondFloor(DateTimeV2Literal date, DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), 1, origin.toJavaDateType(), false));
-    }
-
-    @ExecFunction(name = "second_floor", argTypes = {"DATETIMEV2", "INT", "DATETIMEV2"}, returnType = "DATETIMEV2")
-    public static DateTimeV2Literal secondFloor(DateTimeV2Literal date, IntegerLiteral period,
-            DateTimeV2Literal origin) {
-        return DateTimeV2Literal.fromJavaDateType(
-                getDateFloor(DATE.SECOND, date.toJavaDateType(), period.getValue(), origin.toJavaDateType(), false));
-    }
 
     /**
      * other scalar function
